feat: add notification service and tools
Implement a cross-platform NotificationService and integrate it into
the ChatService. Add support for notifications, scheduling, and
cancellation via new chat tools: notify, list_notifications, and
cancel_notification.

On Windows, make use of native toast notifications to ensure reminders
and alerts are handled even when the app is closed. Introduce a
reference implementation for MacCatalyst, setting a groundwork for
future development.

Ensure the UI is informed of necessary platform interactions and
correctly handles notification requests. Enable Floaty to manage
scheduled notifications effectively, enhancing user interaction and
task management capabilities.

+++ app/Platforms/Windows/WindowsNotificationService.cs
@@ -281,7 +281,7 @@ public sealed class WindowsNotificationService : INotificationService
         try
         {
             var nodes = content.GetElementsByTagName("text");
-            return index < (int)nodes.Count ? nodes[(uint)index].InnerText ?? string.Empty : string.Empty;
+            return index < (int)nodes.Count ? nodes[index].InnerText ?? string.Empty : string.Empty;
         }
         catch
         {
